game: game->utf-8 working for korean, just have to do inverse now
