+++ common/util/font/dbs/font_db_jak1.h
@@ -1,0 +1,25 @@
+#pragma once
+
+#include "common/util/font/font_utils.h"
+
+/*!
+ * ===========================
+ * GAME TEXT FONT BANK - JAK 1
+ * ===========================
+ * This font is used in:
+ * - Jak & Daxter: The Precursor Legacy (Black Label)
+ */
+
+extern GameTextFontBank g_font_bank_jak1_v1;
+/*!
+ * ================================
+ * GAME TEXT FONT BANK - JAK 1 (v2)
+ * ================================
+ * This font is used in:
+ * - Jak & Daxter: The Precursor Legacy (PAL)
+ * - ジャックＸダクスター　～　旧世界の遺産
+ * - Jak & Daxter: The Precursor Legacy (NTSC-U v2)
+ *
+ * It is the same as v1, but _ has been fixed and no longer overlaps 掘
+ */
+extern GameTextFontBank g_font_bank_jak1_v2;
