+++ common/util/font/dbs/font_db_jak2.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "common/util/font/font_utils.h"
+
+extern std::unordered_map<std::string, std::vector<std::string>> jamo_glyph_mappings_jak2;
+
+/*!
+ * ================================
+ * GAME TEXT FONT BANK - JAK 2
+ * ================================
+ * This font is used in:
+ * - Jak II
+ * - Jak II: Renegade
+ * - ジャックＸダクスター2
+ */
+extern GameTextFontBank g_font_bank_jak2;
