+++ common/util/font/dbs/font_db_jak3.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "common/util/font/font_utils.h"
+
+/*!
+ * ================================
+ * GAME TEXT FONT BANK - JAK 3
+ * ================================
+ * This font is used in:
+ * - Jak 3
+ */
+// TODO cyrillic
+// TODO - current just using the jak 2 stuff (hopefully it's identical?)
+extern GameTextFontBank g_font_bank_jak3;
