+++ decompiler/data/game_text.cpp
@@ -9,7 +9,7 @@
 
 #include "common/goos/Reader.h"
 #include "common/util/BitUtils.h"
-#include "common/util/FontUtils.h"
+#include "common/util/font/font_utils.h"
 #include "common/util/print_float.h"
 
 #include "decompiler/ObjectFile/ObjectFileDB.h"
@@ -113,9 +113,9 @@ GameTextResult process_game_text(ObjectFileData& data, GameTextVersion version) 
     }
 
     // escape characters
-    // TODO - hack, skip korean for now, i want the raw bytes
     if (language == 6) {
-      result.text[text_id] = goos::get_byte_string(text.c_str());
+      // If we are doing korean, we process it differently
+      result.text[text_id] = get_font_bank(version)->convert_korean_game_to_utf8(text.c_str());
     } else if (font_bank_exists(version)) {
       result.text[text_id] = get_font_bank(version)->convert_game_to_utf8(text.c_str());
     } else {
