--- game/assets/jak2/temp-text-id-mapping.json
@@ -1,53 +1,0 @@
-{
-  "1a3": "138",
-  "1b5": "161",
-  "133a": "1113",
-  "1200": "1031",
-  "1201": "1032",
-  "1202": "1034",
-  "1203": "1033",
-  "120c": "10c0",
-  "120d": "103b",
-  "120e": "1042",
-  "1215": "10d0",
-  "123a": "1035",
-  "123c": "1070",
-  "123d": "1071",
-  "123f": "107b",
-  "1242": "107c",
-  "1240": "107d",
-  "1241": "107e",
-  "1248": "1050",
-  "1249": "1051",
-  "124e": "100c",
-  "124d": "100e",
-  "124f": "100d",
-  "125e": "1097",
-  "1280": "1000",
-  "1286": "1001",
-  "1287": "1002",
-  "1289": "1030",
-  "128f": "100f",
-  "1290": "1500",
-  "1291": "1600",
-  "1292": "1007",
-  "1293": "160d",
-  "1294": "1601",
-  "1295": "1602",
-  "1296": "1603",
-  "1297": "1604",
-  "1299": "160e",
-  "129a": "160f",
-  "129b": "1610",
-  "129c": "1606",
-  "129d": "1607",
-  "129e": "1608",
-  "129f": "1609",
-  "1300": "1613",
-  "1301": "1611",
-  "1302": "160a",
-  "1303": "160b",
-  "1304": "160c",
-  "1307": "1614",
-  "1308": "1615"
-}
