+++ game/kernel/common/kmachine.cpp
@@ -6,7 +6,7 @@
 #include "common/log/log.h"
 #include "common/symbols.h"
 #include "common/util/FileUtil.h"
-#include "common/util/FontUtils.h"
+#include "common/util/font/font_utils.h"
 #include "common/util/Timer.h"
 #include "common/util/string_util.h"
