+++ game/kernel/jak2/kmachine_extras.cpp
@@ -6,10 +6,9 @@
 #include "kscheme.h"
 
 #include "common/symbols.h"
-#include "common/util/FontUtils.h"
+#include "common/util/font/font_utils.h"
 
 #include "game/external/discord.h"
-#include "game/external/discord_jak1.h"
 #include "game/external/discord_jak2.h"
 #include "game/kernel/common/Symbol4.h"
 #include "game/kernel/common/kmachine.h"
